Fix for Taliban AA infantry, AI now use Stinger
With "CFP_O_TBAN_AA_Specialist_01", if controlled by the AI and orbat randomisation / loadout is applied, the AI unit cannot fire the FIM-92 Stinger .
Fix was to add the Stinger magazine "CUP_Stinger_M" to the loadout. 

Note: there is no issue with a player firing the FIM-92 Stinger if it is in the players control (e.g. launcher picked up by the player, or if the unit is controlled by the player), it only appears to affect the AI's ability to fire.